+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changes:
 
 - Improve Stocard importer
 - Fix importing Catima export with multiline note
+- Scale card title in acceptable range
 
 ## v2.2.0 (2021-08-02)